==> 2. Cox hazards regression.Rmd <==
---
title: "Cox regression for quarterly data"
author: "Bo Havinga"
output: html_document
---


###### --
# QUARTER
###### --

```{r}
library(tidyr)
library(dplyr)
```


Import data
```{r}
KPIdatequarter <- read.csv("/Users/Bo/Documents/Universiteit RSM/Master/Thesis/Data/KPIdatequarter.csv")
KPIdatequarter$apps_MAU <- KPIdatequarter$active_apps/KPIdatequarter$MAU
KPIdatequarter$apps_MAU[is.nan(KPIdatequarter$apps_MAU) | is.infinite(KPIdatequarter$apps_MAU)] <- 0
KPIdatequarter <- KPIdatequarter %>% group_by(ID) %>% mutate(apps_MAU_lag = dplyr::lag(apps_MAU, n=1, default = NA))
```

Delete all rows before start date
```{r}
#install.packages("dplyr")
library(dplyr)
KPIdatequarter <- KPIdatequarter %>%
  filter(tstop != 0)
```


Make event again in case the deleted ones still bought within time period
Make event variable (status)
```{r}
KPIdatequarter$event <- 0
for (i in 1:nrow(KPIdatequarter)) {
  current_tpid <- KPIdatequarter$ID[i]
  current_value <- KPIdatequarter$revenue[i]
  
  previous_value <- KPIdatequarter$revenue[1:(i-1)][KPIdatequarter$ID[1:(i-1)] == current_tpid]
 
  if (current_value >= 10000 & all(previous_value < 10000)) {
    KPIdatequarter$event[i] <- 1
  }
}


```

Make buy variable (churn)
```{r}
KPIdatequarter$buy <- 0
for (i in 1:nrow(KPIdatequarter)) {
  current_id <- KPIdatequarter$ID[i]
  current_event <- KPIdatequarter$event[i]

  if (current_event == 1) {
    KPIdatequarter$buy[KPIdatequarter$ID == current_id] <- 1
  } else if (current_event == 0 & any(KPIdatequarter$buy[KPIdatequarter$ID == current_id] == 1)) {
    KPIdatequarter$buy[i] <- 1
  }
}

```


Delete all rows after event == 1 and all before start date
```{r}
# Sort data by tstop in descending order for each ID
BUYquarterNOTlag <- KPIdatequarter %>%
  arrange(ID, (tstop))

# Filter data based on the criteria you described
BUYquarterNOTlag <- BUYquarterNOTlag %>%
  group_by(ID) %>%
  filter(sum(event) == 0 | event == 1 | (cumsum(event)==0 & sum(event) == 1)) %>%
  ungroup()
```

Make lagged dataset
```{r}
BUYquarter <- BUYquarterNOTlag
# Delete all rows with NA for lagged variables
BUYquarter <- BUYquarter %>%
  group_by(ID) %>%
  filter(!any(is.na(trails_lag)))
```

Save the dataset
```{r}
#write.csv(BUYquarter, file = "/Users/Bo/Documents/Universiteit RSM/Master/Thesis/Data/BUYquarter.csv", row.names = FALSE)
#write.csv(BUYquarterNOTlag, file = "/Users/Bo/Documents/Universiteit RSM/Master/Thesis/Data/BUYquarterNOTlag.csv", row.names = FALSE)
```


```{r}
# When do they buy? 
test <- subset(BUYquarter, buy=="1",
 select =c(tstop, buy) ) 
summary(test) 
print("------Quantile-------")
quantile(test$tstop)
```



```{r}
# Plot when customers buy
library(ggplot2)
ggplot(test, aes(x = tstop)) +
  stat_ecdf(geom = "step", size = 1) +
  labs(x = "Time (in quarters)", y = "Cumulative probability") +
  theme_minimal() +
  theme(
    axis.title.x = element_text(size = 14, margin = margin(t = 10)),
    axis.title.y = element_text(size = 14, margin = margin(r = 10)),
    axis.text = element_text(size = 14)) +
  geom_hline(yintercept = 0.75, linetype = "dashed", color = "red", size = 0.5)
```



```{r}
# Estimation of Survival function using Kaplan-Meier method
#install.packages("survival")
library(survival)
KM_SurvivalCurve <- survfit(Surv(tstop,buy)~ 1, data=BUYquarter) 

#install.packages("survminer")
library(survminer)
ggsurvplot(KM_SurvivalCurve, data=BUYquarter) +
labs(x = "Time (in quarters)", y = "Cumulative probability")
```

Country
```{r}
# Estimation of Survival function using Kaplan-Meier method, for each country
KM_SurvivalCurve2 <- survfit(Surv(tstop,buy)~ country, data=BUYquarter) 
ggsurvplot(KM_SurvivalCurve2, data=BUYquarter, pval = TRUE, ggtheme = theme_void(base_size = 8)) 
```

Industry
```{r}
# Estimation of Survival function using Kaplan-Meier method, for each industry
KM_SurvivalCurve2 <- survfit(Surv(tstop,buy)~ Industry, data=BUYquarter) 
ggsurvplot(KM_SurvivalCurve2, data=BUYquarter, pval = TRUE, ggtheme = theme_void(base_size = 7)) 
```

Segment
```{r}
# Estimation of Survival function using Kaplan-Meier method, for each  segment
KM_SurvivalCurve2 <- survfit(Surv(tstop,buy)~ segment, data=BUYquarter) 
ggsurvplot(KM_SurvivalCurve2, data=BUYquarter, pval = TRUE, ggtheme = theme_void(base_size = 8)) 
```

Make standadized variables
```{r}
BUYquarter$Employees_s <- scale(BUYquarter$Employees)
BUYquarter$cross_selling_s <- scale(BUYquarter$cross_selling)
BUYquarter$apps_MAU_s <- scale(BUYquarter$apps_MAU)
BUYquarter$Avg.Interactions.per.Contact_s <- scale(BUYquarter$Avg.Interactions.per.Contact)
BUYquarter$Avg.Interactions.per.Contact_lag_s <- scale(BUYquarter$Avg.Interactions.per.Contact_lag)

BUYquarter$digital_content_s <- scale(BUYquarter$digital_content)
BUYquarter$events_s <- scale(BUYquarter$events)
BUYquarter$trails_s <- scale(BUYquarter$trails)
BUYquarter$freemium_s <- scale(BUYquarter$freemium)

BUYquarter$digital_content_s_lag <- scale(BUYquarter$digital_content_lag)
BUYquarter$events_s_lag <- scale(BUYquarter$events_lag)
BUYquarter$trails_s_lag <- scale(BUYquarter$trails_lag)
BUYquarter$freemium_s_lag <- scale(BUYquarter$freemium_lag)
```



# Correlation matrix
```{r}
BUYquartercor <- BUYquarter[,c(49,51,50,52,54:57)]
BUYquartercorboth <- BUYquarter[,c(49,51,50,52,54,58,55,59,56,60,57,61)]

# Rename columns
names(BUYquartercor)[1] = "employees"
names(BUYquartercor)[2] = "avg_apps_built"
names(BUYquartercor)[3] = "cross_selling"
names(BUYquartercor)[4] = "avg_interactions_per_employee"
names(BUYquartercor)[5] = "digital_content_downloads"
names(BUYquartercor)[6] = "event_attendees"
names(BUYquartercor)[7] = "requested_trials"
names(BUYquartercor)[8] = "freemium_users"

names(BUYquartercorboth)[1] = "employees"
names(BUYquartercorboth)[2] = "avg_apps_built"
names(BUYquartercorboth)[3] = "cross_selling"
names(BUYquartercorboth)[4] = "avg_interactions_per_employee"
names(BUYquartercorboth)[5] = "digital_content_downloads"
names(BUYquartercorboth)[6] = "digital_content_downloads_lag"
names(BUYquartercorboth)[7] = "event_attendees"
names(BUYquartercorboth)[8] = "event_attendees_lag"
names(BUYquartercorboth)[9] = "requested_trials"
names(BUYquartercorboth)[10] = "requested_trials_lag"
names(BUYquartercorboth)[11] = "freemium_users"
names(BUYquartercorboth)[12] = "freemium_users_lag"

# Make plot
library(corrplot)
corrplot(cor(BUYquartercor), type = "lower",  tl.col = "black", tl.srt = 5, tl.cex = 0.7, cl.cex = 0.7, number.cex = 0.7, addCoef.col = "black")
corrplot(cor(BUYquartercorboth), type = "lower",  tl.col = "black", tl.srt = 5, tl.cex = 0.7, cl.cex = 0.7, number.cex = 0.7, addCoef.col = "black")
```


# -------------------
#Cox regression models
# -------------------

# Null model (not lagged)
```{r}
nullmodel <- coxph(Surv(tstop, event) ~ 1, data=BUYquarter) 

summary(nullmodel) 
print(nullmodel$concordance) 
AIC(nullmodel)                             
```

# Firmographics
```{r}
coxcontrol <- coxph(Surv(tstop, event) ~ Industry + country + segment,
data =BUYquarter, cluster = ID)

summary(coxcontrol)
print(coxcontrol$concordance)
car::vif(coxcontrol)
AIC(coxcontrol) 
```

# Firmographics + controls
```{r}
coxemployeesuse <- coxph(Surv(tstart, tstop, event) ~ Industry + country + segment + Employees_s + apps_MAU_s,
data =BUYquarter, cluster = ID)

summary(coxemployeesuse)
print(coxemployeesuse$concordance) 
car::vif(coxemployeesuse)
AIC(coxemployeesuse)

# I need to use tstart,tstop if the model includes time varying variables (https://cran.r-project.org/web/packages/survival/vignettes/timedep.pdf)
```

# Firmographics + controls + crossselling
```{r}
coxemployeesusecross <- coxph(Surv(tstart, tstop, event) ~ Industry + country + segment + Employees_s + apps_MAU_s + cross_selling_s:tstop,
data =BUYquarter, cluster = ID)

summary(coxemployeesusecross)
print(coxemployeesusecross$concordance)
car::vif(coxemployeesusecross)
AIC(coxemployeesusecross) 
```

# Firmographics + controls + crossselling + marketing
```{r}
coxemployeesusecrossmar <- coxph(Surv(tstart, tstop, event) ~ Industry + country + segment + Employees_s + apps_MAU_s + cross_selling_s:tstop + Avg.Interactions.per.Contact_s:tstop + digital_content_s:tstop + events_s:tstop + trails_s:tstop +freemium_s:tstop,
data =BUYquarter, cluster = ID)

summary(coxemployeesusecrossmar)
print(coxemployeesusecrossmar$concordance)
car::vif(coxemployeesusecrossmar)
AIC(coxemployeesusecrossmar)
```


# Check for Hazards assumption
```{r}
cox.zph(coxemployeesusecrossmar)
```

# Modified after testing assumption
```{r}
coxemployeesusecrossmarstra <- coxph(Surv(tstart, tstop, event) ~ strata(Industry) + country + segment + Employees_s + apps_MAU_s + cross_selling_s:tstop + Avg.Interactions.per.Contact_s:tstop + digital_content_s:tstop + events_s:tstop + trails_s:tstop +freemium_s:tstop,
data =BUYquarter, cluster = ID)

summary(coxemployeesusecrossmarstra)
print(coxemployeesusecrossmarstra$concordance)
car::vif(coxemployeesusecrossmarstra)
AIC(coxemployeesusecrossmarstra)
```

# Test again
```{r}
cox.zph(coxemployeesusecrossmar)
```


# Adding lagged variables
# Wihtout freemium_lag because of high VIF 
```{r}
coxemployeesusecrossmarLAG <- coxph(Surv(tstart, tstop, event) ~ strata(Industry) + country + segment + Employees_s +  cross_selling_s:tstop + apps_MAU_s + Avg.Interactions.per.Contact_s:tstop + digital_content_s:tstop + events_s:tstop + trails_s:tstop +freemium_s:tstop + Avg.Interactions.per.Contact_lag_s:tstop + digital_content_s_lag:tstop + events_s_lag:tstop + trails_s_lag:tstop,
data =BUYquarter, cluster = ID)


summary(coxemployeesusecrossmarLAG)
print(coxemployeesusecrossmarLAG$concordance) 
car::vif(coxemployeesusecrossmarLAG) 
AIC(coxemployeesusecrossmarLAG) 
cox.zph(coxemployeesusecrossmarLAG)
```

# Make nice table
```{r}
aic_values <- c(AIC(nullmodel),
               AIC(coxcontrol),
               AIC(coxemployeesuse),
               AIC(coxemployeesusecross),
               AIC(coxemployeesusecrossmar),
               AIC(coxemployeesusecrossmarstra),
               AIC(coxemployeesusecrossmarLAG))

bic_values <- c(BIC(nullmodel),
               BIC(coxcontrol),
               BIC(coxemployeesuse),
               BIC(coxemployeesusecross),
               BIC(coxemployeesusecrossmar),
               BIC(coxemployeesusecrossmarstra),
               BIC(coxemployeesusecrossmarLAG))

concordance_values <- c(nullmodel$concordance[6],
                        coxcontrol$concordance[6],
                        coxemployeesuse$concordance[6],
                        coxemployeesusecross$concordance[6],
                        coxemployeesusecrossmar$concordance[6],
                        coxemployeesusecrossmarstra$concordance[6],
                        coxemployeesusecrossmarLAG$concordance[6])

# Create the table
model_table <- data.frame(Model = c("Model 1: Null Model",
                                    "Model 2: + Firmographics",
                                    "Model 3: + Control",
                                    "Model 4: + Cross-selling",
                                    "Model 5: + Marketing",
                                    "Model 6: Stratified",
                                    "Model 7: + Lagged"),
                          AIC = aic_values,
                          BIC = bic_values,
                          Concordance = concordance_values)

# Print the table
print(model_table)
latexoutput <- xtable(model_table)
print(latexoutput)
```

# Make Hazard Ratio table
```{r}
# Store the variable names
variables <- c(rownames(summary(coxemployeesusecrossmar)$coefficients), "tstop:Avg.Interactions.per.Contact_lag_s","tstop:digital_content_s_lag","tstop:events_s_lag","tstop:trails_s_lag")
output_table <- data.frame(Variable = variables, stringsAsFactors = FALSE)

# Firmographic
variable_names <- row.names(summary(coxcontrol)$coefficients)
values <- summary(coxcontrol)$coefficients[,2]
df <- data.frame(Variable = variable_names, Value = values)
output_table <- merge(output_table, df, by="Variable", all=TRUE)

# Control
variable_names <- row.names(summary(coxemployeesuse)$coefficients)
values <- summary(coxemployeesuse)$coefficients[,2]
df <- data.frame(Variable = variable_names, Value = values)
output_table <- merge(output_table, df, by="Variable", all=TRUE)

# Cross-selling
variable_names <- row.names(summary(coxemployeesusecross)$coefficients)
values <- summary(coxemployeesusecross)$coefficients[,2]
df <- data.frame(Variable = variable_names, Value = values)
output_table <- merge(output_table, df, by="Variable", all=TRUE)

# Marketing
variable_names <- row.names(summary(coxemployeesusecrossmar)$coefficients)
values <- summary(coxemployeesusecrossmar)$coefficients[,2]
df <- data.frame(Variable = variable_names, Value = values)
output_table <- merge(output_table, df, by="Variable", all=TRUE)

# Stratified
variable_names <- row.names(summary(coxemployeesusecrossmarstra)$coefficients)
values <- summary(coxemployeesusecrossmarstra)$coefficients[,2]
df <- data.frame(Variable = variable_names, Value = values)
output_table <- merge(output_table, df, by="Variable", all=TRUE)

# Lagged
variable_names <- row.names(summary(coxemployeesusecrossmarLAG)$coefficients)
values <- summary(coxemployeesusecrossmarLAG)$coefficients[,2]
df <- data.frame(Variable = variable_names, Value = values)
output_table <- merge(output_table, df, by="Variable", all=TRUE)

names(output_table)[2] = "Model 2: + Firmographics"
names(output_table)[3] = "Model 3: + Control"
names(output_table)[4] = "Model 4: + Cross-selling"
names(output_table)[5] = "Model 5: + Marketing"
names(output_table)[6] = "Model 6: Stratified"
names(output_table)[7] = "Model 7: + Lagged"

latexoutputCOX <- xtable(output_table)
print(latexoutputCOX)
```

```{r}
# Store the variable names
#variables <- c(rownames(summary(coxemployeesusecrossmar)$coefficients), "tstop:Avg.Interactions.per.Contact_lag_s","tstop:digital_content_s_lag","tstop:events_s_lag","tstop:trails_s_lag")
output_table1 <- data.frame(Variable = variables, stringsAsFactors = FALSE)

# Firmographic
variable_names <- row.names(summary(coxcontrol)$coefficients)
values <- summary(coxcontrol)$coefficients[,6]
df <- data.frame(Variable = variable_names, Value = values)
output_table1 <- merge(output_table1, df, by="Variable", all=TRUE)

# Control
variable_names <- row.names(summary(coxemployeesuse)$coefficients)
values <- summary(coxemployeesuse)$coefficients[,6]
df <- data.frame(Variable = variable_names, Value = values)
output_table1 <- merge(output_table1, df, by="Variable", all=TRUE)

# Cross-selling
variable_names <- row.names(summary(coxemployeesusecross)$coefficients)
values <- summary(coxemployeesusecross)$coefficients[,6]
df <- data.frame(Variable = variable_names, Value = values)
output_table1 <- merge(output_table1, df, by="Variable", all=TRUE)

# Marketing
variable_names <- row.names(summary(coxemployeesusecrossmar)$coefficients)
values <- summary(coxemployeesusecrossmar)$coefficients[,6]
df <- data.frame(Variable = variable_names, Value = values)
output_table1 <- merge(output_table1, df, by="Variable", all=TRUE)

# Stratified
variable_names <- row.names(summary(coxemployeesusecrossmarstra)$coefficients)
values <- summary(coxemployeesusecrossmarstra)$coefficients[,6]
df <- data.frame(Variable = variable_names, Value = values)
output_table1 <- merge(output_table1, df, by="Variable", all=TRUE)

# Lagged
variable_names <- row.names(summary(coxemployeesusecrossmarLAG)$coefficients)
values <- summary(coxemployeesusecrossmarLAG)$coefficients[,6]
df <- data.frame(Variable = variable_names, Value = values)
output_table1 <- merge(output_table1, df, by="Variable", all=TRUE)

names(output_table1)[2] = "Model 2: + Firmographics"
names(output_table1)[3] = "Model 3: + Control"
names(output_table1)[4] = "Model 4: + Cross-selling"
names(output_table1)[5] = "Model 5: + Marketing"
names(output_table1)[6] = "Model 6: Stratified"
names(output_table1)[7] = "Model 7: + Lagged"

output_table2 <- output_table1[, 2:7]
output_table2[output_table2 > 0.1] <- NA
output_table2[output_table2 <= 0.01] <- 3
output_table2[output_table2 <= 0.05] <- 2
output_table2[output_table2 <= 0.1] <- 1
output_table2 <- data.frame(variable = output_table1[, 1], output_table2)
```
